Add files via upload
DASHBOARD_CREATION:

A dashboard, on the other hand, is a collection of reports and visualizations that provide a high-level view of the data, allowing users to monitor key metrics and identify trends or patterns. Dashboards often include interactive features such as filters and drill-downs, which allow users to explore the data in more detail.